Add StartPageContentItem
This is currently identical to FlowContentItem. It'll be used to map
between the start page of a Smart Answer and it's representation as
required by the Publishing API.

I've also copied the tests for the FlowContentItem in preparation for
changing the behaviour.

Note the change to the ContentItemPublisher test that now asserts we're
making two requets to put the content and two requests to publish the
content.