Rename Flow#content_id to #start_page_content_id
To make it explicit that this is the content id of the start page. This
will become relevant when we introduce a separate content id for the
Smart Answer flows.

I used `bundle exec rake checksums:update` to update the checksums of
each of the Smart Answer flows. I haven't run the integration tests but
I'm confident that my change won't have broken anything.

This change highlights the fact that we send the Content ID of the Start
Page to Rummager (in the `SearchPayloadPresenter`). This is probably OK
all the time we're publishing Smart Answers (start pages and flows) from
this app but we'll have to work out what to do once we move the start
page content to Publisher.